move authorization logic for websocket
